+++ CHANGELOG.md
@@ -5,6 +5,39 @@
 
 > All notable changes to this project will be documented in this file
 
+## [1.5.0-beta.3](https://github.com/open-sauced/pizza-cli/compare/v1.5.0-beta.2...v1.5.0-beta.3) (2024-09-16)
+
+
+### 🍕 Features
+
+* Cut 2.0.0 release ([#193](https://github.com/open-sauced/pizza-cli/issues/193)) ([278a833](https://github.com/open-sauced/pizza-cli/commit/278a8339c701be1f25db9e3e039f727927b4cdef))
+
+## [1.5.0-beta.2](https://github.com/open-sauced/pizza-cli/compare/v1.5.0-beta.1...v1.5.0-beta.2) (2024-09-13)
+
+
+### 🍕 Features
+
+* Update README docs ([#186](https://github.com/open-sauced/pizza-cli/issues/186)) ([99328aa](https://github.com/open-sauced/pizza-cli/commit/99328aaa322b6dc4abe3f1933b6e77ffedecb355))
+
+## [1.5.0-beta.1](https://github.com/open-sauced/pizza-cli/compare/v1.4.1-beta.1...v1.5.0-beta.1) (2024-09-12)
+
+
+### 🍕 Features
+
+* pizza generate insight command ([#179](https://github.com/open-sauced/pizza-cli/issues/179)) ([7315a1d](https://github.com/open-sauced/pizza-cli/commit/7315a1d2ae2764dcbdfd1ce0af1272fde36b1a0c))
+
+
+### 🐛 Bug Fixes
+
+* Config path prefers local dir vs. home dir ([#184](https://github.com/open-sauced/pizza-cli/issues/184)) ([859446a](https://github.com/open-sauced/pizza-cli/commit/859446ac865e7a4c7d7779a2620c7391b4aabf33))
+
+## [1.4.1-beta.1](https://github.com/open-sauced/pizza-cli/compare/v1.4.0...v1.4.1-beta.1) (2024-09-12)
+
+
+### 🐛 Bug Fixes
+
+* Don't overwrite ldflags in justfile builds ([#171](https://github.com/open-sauced/pizza-cli/issues/171)) ([e024687](https://github.com/open-sauced/pizza-cli/commit/e0246879d7d72095f30305cf805ad86e2df1f623))
+
 ## [1.4.0](https://github.com/open-sauced/pizza-cli/compare/v1.3.0...v1.4.0) (2024-09-11)